Add exception handling

+++ EmployeeWebApp.Application/Companies/DeleteCompany/DeleteCompanyCommandHandler.cs
@@ -1,3 +1,4 @@
+using EmployeeWebApp.Domain.Exceptions;
 using EmployeeWebApp.Domain.Interfaces;
 using MediatR;
 
@@ -22,6 +23,11 @@ public class DeleteCompanyCommandHandler : IRequestHandler<DeleteCompanyCommand>
     /// <param name="cancellationToken"><see cref="CancellationToken"/>.</param>
     public async Task Handle(DeleteCompanyCommand request, CancellationToken cancellationToken = default)
     {
+        if (!await _companyRepository.CompanyExistsAsync(request.Id))
+        {
+            throw new CompanyNotFoundException(request.Id);
+        }
+        
         await _companyRepository.DeleteCompanyAsync(request.Id);
     }
 }
